Conflicts:
	run/pass_gen.pl
	src/AFS_fmt.c
	src/BSDI_fmt.c
	src/DES_std.c
	src/DES_std.h
	src/EPI_fmt_plug.c
	src/KRB5_fmt_plug.c
	src/LM_fmt.c
	src/MSCHAPv2_fmt_plug.c
	src/NETLM_fmt_plug.c
	src/NETLMv2_fmt_plug.c
	src/NETNTLM_fmt_plug.c
	src/NETNTLMv2_fmt_plug.c
	src/NETSPLITLM_fmt_plug.c
	src/PHPS_fmt_plug.c
	src/cuda_cryptmd5_fmt.c
	src/cuda_cryptsha256_fmt.c
	src/cuda_cryptsha512_fmt.c
	src/cuda_mscash2_fmt.c
	src/cuda_mscash_fmt.c
	src/cuda_phpass_fmt.c
	src/cuda_pwsafe_fmt.c
	src/cuda_rawsha256_fmt.c
	src/cuda_wpapsk_fmt.c
	src/dynamic_fmt.c
	src/dynamic_preloads.c
	src/formats.c
	src/formats.h
	src/mediawiki_fmt_plug.c
	src/opencl_cryptmd5_fmt.c
	src/opencl_cryptsha512_fmt.c
	src/opencl_mscash2_fmt.c
	src/opencl_nsldaps_fmt.c
	src/opencl_phpass_fmt.c
	src/opencl_pwsafe_fmt.c
	src/opencl_wpapsk_fmt.c
	src/osc_fmt_plug.c
	src/phpassMD5_fmt_plug.c
	src/pixMD5_fmt_plug.c
	src/rawSHA1_linkedIn_fmt_plug.c
	src/rawSHA512_fmt_plug.c
	src/trip_fmt.c
	src/unused/rawMD5_thin_fmt_plug.c
	src/unused/rawmd5u_thin_fmt_plug.c

+++ src/dummy.c
@@ -80,7 +80,7 @@ static int valid(char *ciphertext, struct fmt_main *self)
 			warned = 2;
 			fprintf(stderr,
 			        "dummy password length %d > max. supported lengh %d\n",
-				((q - p) >> 1), MAX_PLAINTEXT_LENGTH);
+				(int)((q - p) >> 1), MAX_PLAINTEXT_LENGTH);
 		}
 		else if (warned == 0 && ((q - p) >> 1) > PLAINTEXT_LENGTH) {
 			warned = 1;
@@ -90,7 +90,7 @@ static int valid(char *ciphertext, struct fmt_main *self)
 			 */
 			fprintf(stderr,
 			        "dummy password length %d > currently supported length %d\n",
-			        ((q - p) >> 1), PLAINTEXT_LENGTH);
+			        (int)((q - p) >> 1), PLAINTEXT_LENGTH);
 		}
 		return 0;
 	}
